// refatoring/frontend2/src/gen/api/src/models/SelectOptionDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SelectOptionDto
 */
export interface SelectOptionDto {
    /**
     * 
     * @type {string}
     * @memberof SelectOptionDto
     */
    name?: string | null;
    /**
     * 
     * @type {number}
     * @memberof SelectOptionDto
     */
    value?: number;
}

/**
 * Check if a given object implements the SelectOptionDto interface.
 */
export function instanceOfSelectOptionDto(value: object): value is SelectOptionDto {
    return true;
}

export function SelectOptionDtoFromJSON(json: any): SelectOptionDto {
    return SelectOptionDtoFromJSONTyped(json, false);
}

export function SelectOptionDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SelectOptionDto {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'value': json['value'] == null ? undefined : json['value'],
    };
}

export function SelectOptionDtoToJSON(json: any): SelectOptionDto {
    return SelectOptionDtoToJSONTyped(json, false);
}

export function SelectOptionDtoToJSONTyped(value?: SelectOptionDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'value': value['value'],
    };
}