// refatoring/frontend2/src/gen/api/src/models/UserInfoResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { EActivityLocation } from './EActivityLocation';
import {
    EActivityLocationFromJSON,
    EActivityLocationFromJSONTyped,
    EActivityLocationToJSON,
    EActivityLocationToJSONTyped,
} from './EActivityLocation';

/**
 * 
 * @export
 * @interface UserInfoResponseDto
 */
export interface UserInfoResponseDto {
    /**
     * 
     * @type {number}
     * @memberof UserInfoResponseDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof UserInfoResponseDto
     */
    email?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserInfoResponseDto
     */
    nickname?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserInfoResponseDto
     */
    password?: string | null;
    /**
     * 
     * @type {EActivityLocation}
     * @memberof UserInfoResponseDto
     */
    activityLocation?: EActivityLocation;
    /**
     * 
     * @type {string}
     * @memberof UserInfoResponseDto
     */
    profilePicture?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserInfoResponseDto
     */
    profileBanner?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UserInfoResponseDto
     */
    bannerPosition?: number | null;
}



/**
 * Check if a given object implements the UserInfoResponseDto interface.
 */
export function instanceOfUserInfoResponseDto(value: object): value is UserInfoResponseDto {
    return true;
}

export function UserInfoResponseDtoFromJSON(json: any): UserInfoResponseDto {
    return UserInfoResponseDtoFromJSONTyped(json, false);
}

export function UserInfoResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserInfoResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'email': json['email'] == null ? undefined : json['email'],
        'nickname': json['nickname'] == null ? undefined : json['nickname'],
        'password': json['password'] == null ? undefined : json['password'],
        'activityLocation': json['activityLocation'] == null ? undefined : EActivityLocationFromJSON(json['activityLocation']),
        'profilePicture': json['profilePicture'] == null ? undefined : json['profilePicture'],
        'profileBanner': json['profileBanner'] == null ? undefined : json['profileBanner'],
        'bannerPosition': json['bannerPosition'] == null ? undefined : json['bannerPosition'],
    };
}

export function UserInfoResponseDtoToJSON(json: any): UserInfoResponseDto {
    return UserInfoResponseDtoToJSONTyped(json, false);
}

export function UserInfoResponseDtoToJSONTyped(value?: UserInfoResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'email': value['email'],
        'nickname': value['nickname'],
        'password': value['password'],
        'activityLocation': EActivityLocationToJSON(value['activityLocation']),
        'profilePicture': value['profilePicture'],
        'profileBanner': value['profileBanner'],
        'bannerPosition': value['bannerPosition'],
    };
}